see #120 Split coursecontroller and admincontroller to make it easier to understand coursecontroller for future extension
